separate dukan style to 3 composable styles
